+++ spec/lib/paypal-api/client_spec.rb
@@ -84,6 +84,7 @@ RSpec.describe PaypalAPI::Client do
       shipment_tracking: PaypalAPI::ShipmentTracking,
       subscriptions: PaypalAPI::Subscriptions,
       subscription_plans: PaypalAPI::SubscriptionPlans,
+      transaction_search: PaypalAPI::TransactionSearch,
       user_info: PaypalAPI::UserInfo,
       users: PaypalAPI::Users,
       webhooks: PaypalAPI::Webhooks,