+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add PartnerReferrals APIs collection
 - Add PaymentExperienceWebProfiles APIs collection
+- Add TransactionSearch APIs collection
 
 ## [0.4.0] - 2024-09-16